feat: Implement fuction get_universe
Creates a stock universe by selecting only those companies that have a market capitalization of at least 1 billion dollars OR that are in the previous day's holdings, even if on the current day, the company no longer meets the 1 billion dollar criteria. Use the .copy() attribute to create a copy of the data and make sure that we are not looking at returns when forming the portfolio! To make this impossible, make sure to drop the column containing the daily return.